requirements.txt
grpcio==1.35.0
grpcio-tools==1.35.0
numpy==1.19.5
